foot race player flag fix (#1125)
and include fixes

+++ dScripts/ai/ACT/FootRace/BaseFootRaceManager.cpp
@@ -6,8 +6,7 @@ void BaseFootRaceManager::OnStartup(Entity* self) {
 	// TODO: Add to FootRaceStarter group
 }
 
-void BaseFootRaceManager::OnFireEventServerSide(Entity* self, Entity* sender, std::string args, int32_t param1,
-	int32_t param2, int32_t param3) {
+void BaseFootRaceManager::OnFireEventServerSide(Entity* self, Entity* sender, std::string args, int32_t param1, int32_t param2, int32_t param3) {
 	const auto splitArguments = GeneralUtils::SplitString(args, '_');
 	if (splitArguments.size() > 1) {
 
@@ -37,7 +36,7 @@ void BaseFootRaceManager::OnFireEventServerSide(Entity* self, Entity* sender, st
 					if (character != nullptr) {
 						character->SetPlayerFlag(115, false);
 						if (param2 != -1) // Certain footraces set a flag
-							character->SetPlayerFlag(static_cast<uint32_t>(param2), true);
+							character->SetPlayerFlag(param2, true);
 					}
 
 					StopActivity(self, player->GetObjectID(), 0, param1);
